adding pairHMM

+++ src/model/allele_frequency_calculator_result.rs
@@ -112,7 +112,7 @@ impl AFCalculationResult {
 
     pub fn passes_threshold(&self, allele: &Allele, phred_scale_qual_threshold: f64) -> bool {
         (self.get_log10_posterior_of_allele_absent(allele) + AFCalculationResult::EPSILON)
-            < QualityUtils::qual_to_error_prob_log10(phred_scale_qual_threshold)
+            < QualityUtils::qual_to_error_prob_log10(phred_scale_qual_threshold as u8)
     }
 
     /**
